# game-lb-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v3/game-service/inner/area-score": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对弈"
                ],
                "summary": "数目",
                "parameters": [
                    {
                        "description": "对弈",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OwnershipRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.InnerAreaScoreResponse"
                        }
                    }
                }
            }
        },
        "/api/v3/game-service/inner/call-end": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内部"
                ],
                "summary": "结束",
                "parameters": [
                    {
                        "description": "对弈",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CallEndRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v3/game-service/inner/init": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "对弈"
                ],
                "summary": "初始化(内部)",
                "parameters": [
                    {
                        "description": "对弈",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.InfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.CallEndRequest": {
            "type": "object",
            "required": [
                "game_id",
                "win",
                "win_result"
            ],
            "properties": {
                "end_time": {
                    "type": "integer"
                },
                "game_id": {
                    "type": "integer"
                },
                "sgf": {
                    "type": "string"
                },
                "start_time": {
                    "type": "integer"
                },
                "win": {
                    "description": "1 黑 2 白 3 和",
                    "type": "integer"
                },
                "win_result": {
                    "description": "输赢原因 B+R 黑胜白认输 B+T 黑胜白超时 W+R 白胜黑认输 W+T 白胜黑超时 B+C1 黑胜吃1子 W+C1 白胜吃一子 B+L 黑胜白未参加 W+L 白胜黑未参加 Draw 和",
                    "type": "string"
                }
            }
        },
...
        "request.InfoRequest": {
            "type": "object",
            "required": [
                "game_id",
                "user_hash",
                "user_id"
            ],
            "properties": {
                "game_id": {
                    "type": "integer"
                },
                "user_hash": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "request.OwnershipRequest": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "integer"
                },
                "sgf": {
                    "type": "string"
                }
            }
        },
...
        "responses.InnerAreaScoreResponse": {
            "type": "object",
            "properties": {
                "b_score": {
                    "type": "number"
                },
                "controversy_count": {
                    "type": "integer"
                },
                "end_score": {
                    "type": "number"
                },
                "w_score": {
                    "type": "number"
                }
            }
        },
...
        "responses.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0 成功 1失败",
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "description": "提示语",
                    "type": "string"
                }
            }
        }
    }
}